removed searchResult page and instead shows search results in explore page.
