+++ client/src/pages/Explore.jsx
@@ -1,13 +1,12 @@
 /**
  * Explore Component
  * The Explore component displays a search form, a list of popular dishes, and a button to search for recipes. 
+ * When a user clicks search button, shows search results. Filter items persist until user clears it.
  * @component
 */
 
 import { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import { LuSettings2 } from 'react-icons/lu';
-// import { FaSearch } from "react-icons/fa";
 import RecipeList from "../components/explore/RecipeList";
 import BackendAPI from "../helper/BackendApi";
 import Loader from "../components/Loader";
@@ -16,12 +15,14 @@ import Filter from "../components/filter/filter";
 
 
 export default function Explore() {
-    const navigate = useNavigate();
     const [randomRecipe, setRandomRecipe] = useState("");
     const [searchTerm, setSearchTerm] = useState("");
     const [error, setError] = useState("");
     const [searchFilter, setSearchFilter] = useState({});
     const [isFilter, setIsFilter] = useState(false);
+    const [searchOn, setSearchOn] = useState(false);
+    const [recipes, setRecipes] = useState([]);
+    const [isLoading, setIsLoading] = useState(false);
 
     useEffect(() => {
         // console.log(`useEffect ran`)
@@ -31,6 +32,32 @@ export default function Explore() {
     useEffect(()=>{
         console.log(searchFilter);
     },[isFilter,searchFilter])
+
+    useEffect(()=>{
+        if(searchTerm.length===0){
+        setSearchOn(false);
+        }},[searchTerm]);
+
+    useEffect(()=>{
+        
+        async function getRecipes(term){
+            console.log(`is this running?`)
+            try{
+                const response = await BackendAPI.searchRecipe(term);
+                // console.log("🚀 ~ file: Explore.jsx:47 ~ getRecipes ~ response:", response)
+                setRecipes(response);
+                setSearchOn(false);
+            } catch(e){
+                console.log('something went wrong inside useEffect', e)
+            }
+            setIsLoading(false);
+        }        
+        setIsLoading(true);
+        if(searchOn){
+            getRecipes({q:searchTerm, ...searchFilter}); 
+        }
+    }, [searchOn])
+
 
     async function getRecipe() {
         try {
@@ -46,7 +73,7 @@ export default function Explore() {
         if (!searchTerm) {
             setError("You cannot search an empty word.");
         } else {
-            navigate(`/result/${searchTerm}`, {state:searchFilter});
+            setSearchOn(true);
         }
     }
 
@@ -74,7 +101,7 @@ export default function Explore() {
                         <form>
                             <label htmlFor="default-search" className="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white">Search</label>
                             <div className="relative">
-                                <input type="search" id="default-search" className="block w-full p-4  text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-green-500 focus:border-green-500" placeholder="What are you craving?" onChange={handleChange} required />
+                                <input type="search" id="default-search" value ={searchTerm} className="block w-full p-4  text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-green-500 focus:border-green-500" placeholder="What are you craving?" onChange={handleChange} required />
                                 <button type="submit" onClick={(e) => handleSubmit(e)} className="absolute top-0 right-0 p-2.5 text-sm font-medium h-full text-white bg-green-700 rounded-r-lg border border-green-700 hover:bg-green-800 focus:ring-4 focus:outline-none focus:ring-green-300">
                                     <svg className="w-4 h-4" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
                                         <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z" />
@@ -86,8 +113,16 @@ export default function Explore() {
                         </form>
                         {error.length > 0 && <Error error={error} />}
                     </section>
-                    <h2 className="font-bold text-xl mt-5">Popular dishes</h2>
-                    <RecipeList recipes={randomRecipe} />
+                    {searchOn
+                    ?(
+                        isLoading? (<Loader />):(<RecipeList recipes={recipes} />)
+                    )
+                    :(
+                        <div>
+                            <h2 className="font-bold text-xl mt-5">Popular dishes</h2>
+                            <RecipeList recipes={randomRecipe} />
+                        </div>
+                    )}
                 </div>
             </main>
     );
